@@ -1,25 +1,46 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
+/**
+ * Class Image
+ * @package App\Models
+ *
+ * @property string name
+ * @property int advert_id
+ */
 class Image extends Model
 {
     use HasFactory;
 
-    protected $fillable = [
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var string[]
+     */
+    protected array $fillable = [
         'name',
         'advert_id',
     ];
 
     /**
+     * Indicates if the model should be timestamped.
+     *
      * @var bool
      */
-    public $timestamps = false;
+    public bool $timestamps = false;
 
-    public function advert () {
+    /**
+     * @return BelongsTo
+     */
+    public function advert(): BelongsTo
+    {
         return $this->belongsTo(Advert::class);
     }
 }
