@@ -6,10 +6,15 @@ namespace App\Models\Locality;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 /**
  * Class City
  * @package App\Models\Locality
+ *
+ * @property int id
+ * @property string name
+ * @property int region_id
  */
 class City extends Model
 {
@@ -18,7 +23,7 @@ class City extends Model
     /**
      * @var string[]
      */
-    protected $fillable = [
+    protected array $fillable = [
         'name',
         'region_id',
     ];
@@ -26,9 +31,13 @@ class City extends Model
     /**
      * @var bool
      */
-    public $timestamps = false;
+    public bool $timestamps = false;
 
-    public function region () {
+    /**
+     * @return BelongsTo
+     */
+    public function region(): BelongsTo
+    {
         return $this->belongsTo(Region::class);
     }
 }
