@@ -12,7 +12,7 @@ class CategoryFactory extends Factory
      *
      * @var string
      */
-    protected $model = Category::class;
+    protected string $model = Category::class;
 
     /**
      * Define the model's default state.
