@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers\Advert;
 
 use App\Http\Controllers\Controller;
@@ -7,11 +9,12 @@ use App\Http\Requests\Advert\CreateRequest;
 use App\Http\Requests\Advert\UpdateRequest;
 use App\Models\Advert;
 use App\Models\Category;
-use App\Models\Image;
 use App\Models\User;
 use App\Services\AdvertService\AdvertService;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Storage;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
 
 /**
  * Class AdvertController
@@ -30,6 +33,9 @@ class AdvertController extends Controller
         $this->service = $service;
     }
 
+    /**
+     * @return Application|Factory|View
+     */
     public function index()
     {
         $adverts = Advert::with(['category', 'user'])
@@ -42,6 +48,9 @@ class AdvertController extends Controller
         return view('advert.index', compact('adverts'));
     }
 
+    /**
+     * @return Application|Factory|View
+     */
     public function create()
     {
         $categories = Category::defaultOrder()->withDepth()->get();
@@ -49,13 +58,21 @@ class AdvertController extends Controller
         return view('advert.create', compact('categories'));
     }
 
-    public function store(CreateRequest $request)
+    /**
+     * @param CreateRequest $request
+     * @return RedirectResponse
+     */
+    public function store(CreateRequest $request): RedirectResponse
     {
         $advert = $this->service->create($request->all());
 
         return redirect()->route('advert.show', $advert);
     }
 
+    /**
+     * @param Advert $advert
+     * @return Application|Factory|View
+     */
     public function show(Advert $advert)
     {
         $advert->timestamps = false;
@@ -64,6 +81,10 @@ class AdvertController extends Controller
         return view('advert.show', compact('advert'));
     }
 
+    /**
+     * @param Advert $advert
+     * @return Application|Factory|View
+     */
     public function edit(Advert $advert)
     {
         $categories = Category::defaultOrder()->withDepth()->get();
@@ -71,6 +92,11 @@ class AdvertController extends Controller
         return view('advert.edit', compact('advert', 'categories'));
     }
 
+    /**
+     * @param UpdateRequest $request
+     * @param Advert $advert
+     * @return Application|Factory|View
+     */
     public function update(UpdateRequest $request, Advert $advert)
     {
         $this->service->update($advert, $request);
@@ -78,7 +104,11 @@ class AdvertController extends Controller
         return view('advert.show', compact('advert'));
     }
 
-    public function destroy(Advert $advert)
+    /**
+     * @param Advert $advert
+     * @return RedirectResponse
+     */
+    public function destroy(Advert $advert): RedirectResponse
     {
         $this->service->destroy($advert);
 
