@@ -1,10 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Providers;
 
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
 
+/**
+ * Class AuthServiceProvider
+ * @package App\Providers
+ */
 class AuthServiceProvider extends ServiceProvider
 {
     /**
@@ -25,6 +31,16 @@ class AuthServiceProvider extends ServiceProvider
     {
         $this->registerPolicies();
 
-        //
+        Gate::define('view-advert', function ($user, $advert) {
+            return $advert->isAuthor($user->id);
+        });
+
+        Gate::define('update-advert', function ($user, $advert) {
+            return $advert->isAuthor($user->id);
+        });
+
+        Gate::define('delete-advert', function ($user, $advert) {
+            return $advert->isAuthor($user->id);
+        });
     }
 }
